requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
python-dotenv==0.17.1
Werkzeug==2.0.1
